// web-application/src/types-openapi/api/models/InventoryCheck.ts
export interface InventoryCheck {
    /**
     * 
     * @type {Date}
     * @memberof InventoryCheck
     */
    createdAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof InventoryCheck
     */
    updatedAt?: Date | null;
    /**
     * 
     * @type {string}
     * @memberof InventoryCheck
     */
    createdBy?: string | null;
    /**
     * 
     * @type {string}
     * @memberof InventoryCheck
     */
    updatedBy?: string | null;
    /**
     * 
     * @type {number}
     * @memberof InventoryCheck
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof InventoryCheck
     */
    code?: string | null;
    /**
     * 
     * @type {Date}
     * @memberof InventoryCheck
     */
    checkTime?: Date;
    /**
     * 
     * @type {InventoryCheckStatus}
     * @memberof InventoryCheck
     */
    status?: InventoryCheckStatus;
    /**
     * 
     * @type {Date}
     * @memberof InventoryCheck
     */
    balancedAt?: Date | null;
    /**
     * 
     * @type {string}
     * @memberof InventoryCheck
     */
    note?: string | null;
    /**
     * 
     * @type {Array<InventoryCheckItem>}
     * @memberof InventoryCheck
     */
    items?: Array<InventoryCheckItem> | null;
}